+++ src/modules/ThatPasteButton.cpp
@@ -12,6 +12,58 @@ bool ThatPasteButton::onSettingChanged(std::string_view key, const matjson::Valu
     return true;
 }
 
+bool ThatPasteButton::isGroupDefault(GameObject* object) {
+    if (object->m_groups) {
+        return object->m_groups->empty();
+    }
+    return true;
+}
+
+bool ThatPasteButton::areObjectGroupsDefault() {
+    if (m_editorUI->m_selectedObject) {
+        return isGroupDefault(m_editorUI->m_selectedObject);
+    }
+    else if (m_editorUI->m_selectedObjects) {
+        for (auto obj : m_editorUI->m_selectedObjects->asExt<GameObject>()) {
+            bool isDefault = isGroupDefault(obj);
+            if (!isDefault) return false;
+        }
+    }
+
+    return true;
+}
+
+bool ThatPasteButton::isColorDefault(GameObject* object) {
+    bool defaultBase = true;
+    if (object->m_baseColor) {
+        defaultBase = object->m_baseColor->m_colorID == 0 || object->m_baseColor->m_colorID == object->m_baseColor->m_defaultColorID;
+    }
+
+    bool defaultDetail = true;
+    if (object->m_detailColor) {
+        defaultDetail = object->m_detailColor->m_colorID == 0 || object->m_detailColor->m_colorID == object->m_detailColor->m_defaultColorID;
+    }
+
+    if (defaultBase && defaultDetail) {
+        return true;
+    }
+    return false;
+}
+
+bool ThatPasteButton::areObjectColorsDefault() {
+    if (m_editorUI->m_selectedObject) {
+        return isColorDefault(m_editorUI->m_selectedObject);
+    }
+    else if (m_editorUI->m_selectedObjects) {
+        for (auto obj : m_editorUI->m_selectedObjects->asExt<GameObject>()) {
+            bool isDefault = isColorDefault(obj);
+            if (!isDefault) return false;
+        }
+    }
+
+    return true;
+}
+
 void ThatPasteButton::onEditor() {
     auto buttons = m_editorUI->getChildByID("editor-buttons-menu");
     if (!buttons) return;
@@ -22,10 +74,14 @@ void ThatPasteButton::onEditor() {
         tinker::utils::hijackButton(pasteStateButton, [this, pasteStateButton] (auto orig, auto sender) {
             if (!ThatPasteButton::isEnabled() || !getSetting<bool, "toggle-paste-state">()) return orig(sender);
 
-            bool alwaysShow = getSetting<bool, "always-show">();
-            bool hasMultiple = m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1;
+            int showCount = getSetting<int, "object-requirement">();
+            bool show = (m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() >= showCount) || (showCount == 1 && m_editorUI->m_selectedObject);
 
-            if (((alwaysShow && m_editorUI->m_selectedObject || hasMultiple) || (!alwaysShow && hasMultiple)) && pasteStateButton->getOpacity() == 255) {
+            if (show && pasteStateButton->getOpacity() == 255) {
+                if (getSetting<bool, "dont-show-if-default">() && areObjectGroupsDefault()) {
+                    orig(sender);
+                    return;
+                }
                 createQuickPopup("Paste State?", "Pasting state is <cr>dangerous</c>! Are you sure?", "Cancel", "Yes", [this, orig, sender] (FLAlertLayer*, bool yes) {
                     if (yes) orig(sender);
                 });
@@ -39,11 +95,15 @@ void ThatPasteButton::onEditor() {
 
         tinker::utils::hijackButton(pasteColorButton, [this, pasteColorButton] (auto orig, auto sender) {
             if (!ThatPasteButton::isEnabled() || !getSetting<bool, "toggle-paste-color">()) return orig(sender);
+
+            int showCount = getSetting<int, "object-requirement">();
+            bool show = (m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() >= showCount) || (showCount == 1 && m_editorUI->m_selectedObject);
             
-            bool alwaysShow = getSetting<bool, "always-show">();
-            bool hasMultiple = m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1;
-            
-            if (((alwaysShow && m_editorUI->m_selectedObject || hasMultiple) || (!alwaysShow && hasMultiple)) && pasteColorButton->getOpacity() == 255) {
+            if (show && pasteColorButton->getOpacity() == 255) {
+                if (getSetting<bool, "dont-show-if-default">() && areObjectColorsDefault()) {
+                    orig(sender);
+                    return;
+                }
                 createQuickPopup("Paste Color?", "Pasting color is <cr>dangerous</c>! Are you sure?", "Cancel", "Yes", [this, orig, sender] (FLAlertLayer*, bool yes) {
                     if (yes) orig(sender);
                 });
@@ -65,10 +125,14 @@ void ThatPasteButton::onSetGroupIDLayer(SetGroupIDLayer* setGroupIDLayer, GameOb
     tinker::utils::hijackButton(pasteButton, [this] (auto orig, auto sender) {
         if (!ThatPasteButton::isEnabled() || !getSetting<bool, "toggle-paste-state-group">()) return orig(sender);
 
-        bool alwaysShow = getSetting<bool, "always-show">();
-        bool hasMultiple = m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1;
-                
-        if (((alwaysShow && m_editorUI->m_selectedObject || hasMultiple) || (!alwaysShow && hasMultiple))) {
+        int showCount = getSetting<int, "object-requirement">();
+        bool show = (m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() >= showCount) || (showCount == 1 && m_editorUI->m_selectedObject);
+            
+        if (show) {
+            if (getSetting<bool, "dont-show-if-default">() && areObjectGroupsDefault()) {
+                orig(sender);
+                return;
+            }
             createQuickPopup("Paste State?", "Pasting state is <cr>dangerous</c>! Are you sure?", "Cancel", "Yes", [this, orig, sender] (FLAlertLayer*, bool yes) {
                 if (yes) orig(sender);
             });